+++ packages/rpc-provider/src/ws/index.ts
@@ -32,7 +32,7 @@ interface WsStateSubscription extends SubscriptionHandler {
   params: any[];
 }
 
-const ALIASSES: { [index: string]: string } = {
+const ALIASES: { [index: string]: string } = {
   chain_finalisedHead: 'chain_finalizedHead',
   chain_subscribeFinalisedHeads: 'chain_subscribeFinalizedHeads',
   chain_unsubscribeFinalisedHeads: 'chain_unsubscribeFinalizedHeads'
@@ -225,7 +225,6 @@ export class WsProvider implements ProviderInterface {
 
       // 1000 - Normal closure; the connection successfully completed
       this.#websocket.close(1000);
-      this.#websocket = null;
     } catch (error) {
       l.error(error);
 
@@ -351,9 +350,19 @@ export class WsProvider implements ProviderInterface {
     }
 
     this.#isConnected = false;
+
+    if (this.#websocket) {
+      this.#websocket.onclose = null;
+      this.#websocket.onerror = null;
+      this.#websocket.onmessage = null;
+      this.#websocket.onopen = null;
+      this.#websocket = null;
+    }
+
     this.#emit('disconnected');
+
     // reject all hanging requests
-    eraseRecord(this.#handlers, (handler) => handler.callback(error, undefined));
+    eraseRecord(this.#handlers, (h) => h.callback(error, undefined));
     eraseRecord(this.#waitingForId);
 
     if (this.#autoConnectMs > 0) {
@@ -419,7 +428,7 @@ export class WsProvider implements ProviderInterface {
   }
 
   #onSocketMessageSubscribe = (response: JsonRpcResponse): void => {
-    const method = ALIASSES[response.method as string] || response.method || 'invalid';
+    const method = ALIASES[response.method as string] || response.method || 'invalid';
     const subId = `${method}::${response.params.subscription}`;
     const handler = this.#subscriptions[subId];