+++ packages/types/src/interfaces/contracts/types.ts
@@ -8,10 +8,10 @@ import type { AccountId, Balance, BlockNumber, Hash, Weight } from '@polkadot/ty
 export interface AliveContractInfo extends Struct {
   readonly trieId: TrieId;
   readonly storageSize: u32;
-  readonly emptyPairCount: u32;
-  readonly totalPairCount: u32;
+  readonly pairCount: u32;
   readonly codeHash: CodeHash;
   readonly rentAllowance: Balance;
+  readonly rentPaid: Balance;
   readonly deductBlock: BlockNumber;
   readonly lastWrite: Option<BlockNumber>;
   readonly _reserved: Option<Null>;