+++ packages/types/src/interfaces/parachains/definitions.ts
@@ -121,7 +121,8 @@ export default {
       availabilityVotes: 'BitVec',
       backers: 'BitVec',
       relayParentNumber: 'BlockNumber',
-      backedInNumber: 'BlockNumber'
+      backedInNumber: 'BlockNumber',
+      backingGroup: 'GroupIndex'
     },
     CandidateReceipt: {
       descriptor: 'CandidateDescriptor',
@@ -310,8 +311,10 @@ export default {
       relayParentStorageRoot: 'Hash',
       maxPovSize: 'u32'
     },
-    RelayChainBlockNumber: 'u32',
-    RelayChainHash: 'Hash',
+    RelayBlockNumber: 'u32',
+    RelayChainBlockNumber: 'RelayBlockNumber',
+    RelayHash: 'Hash',
+    RelayChainHash: 'RelayHash',
     QueuedParathread: {
       claim: 'ParathreadEntry',
       coreOffset: 'u32'